lab1/src/requirements.txt
pandas
matplotlib
seaborn
requests
python-dotenv